feat(lsp): port LSP client/registry/protocol/server_config to dasclaw_lsp (#629)
把 LSP 的核心实现从 desktop-client/ironclaw/src/tools/builtin/lsp/ 搬到新 crate dasclaw_lsp，4 个文件原样搬迁（verbatim port，ADR-129 §1.3）：

- client.rs（JSON-RPC over stdio + 超时握手）

- protocol.rs（LspAction / LspRequest / LspNotification / LspResponse / LspError）

- server_config.rs（rust-analyzer / typescript-language-server / pyright / gopls 默认映射 + admin 白名单过滤）

- mod.rs → registry.rs（LspRegistry 懒启动 + idle 超时回收）

实际改动仅是顶层模块声明从 mod.rs 拆到 lib.rs，并把 `crate::tools::tool::ToolError` 换成 `dasclaw_tool::ToolError`（两条 use 行）。LspQueryTool 暂留在 ironclaw 侧，等后续 issue 抽离 Tool trait 后再搬。ironclaw 侧通过 lsp/mod.rs 的 pub use 做 facade，外部调用点（registry.rs:834）无需改动。

验证：

- cargo check -p dasclaw_lsp --tests: OK

- cargo nextest run -p dasclaw_lsp: 20/20

- cargo check -p dasclaw --tests: OK（2 个 warning 是 dasclaw_mcp 历史遗留，与本次无关）

- cargo nextest run -p dasclaw --lib tools::builtin::lsp: 17/17

- cargo fmt --all: OK

- python3.12 scripts/check_no_panics.py --base origin/xClaw: OK

- python3 scripts/check_no_new_ironclaw_literal.py --base origin/xClaw: OK（类A）

- cargo clippy --no-deps -p dasclaw_lsp --all-targets -- -D warnings: 0 警告

Cross-cuts: 类A（无新增 .ironclaw / IRONCLAW_BASE_DIR 字面量）

Refs: ADR-152 §3 F3.3, Epic #621
