+++ crates/dasclaw_lsp/src/protocol.rs
